Additional Test added for MySqlAdapter + Adapter fixes
Added a specific "ShouldDeleteDataRepectingRelationships" test to
check deletions occur in the intended order when tables are referenced
as we saw failures when running against app database despite all
existing DbAdapter tests passing.

This identified issues with the adapter which this commit also fixes.

The test checks that tables with multiple relationships (both FK and
PK) all manage to delete, this can only happen if the order of deletion
is properly respected.

The scenario is as-follows:

Foo.FooValue (PK)
Foo.BobValue (FK) ==> (Bob.BobValue)

Bob.BobValue (PK)

Bar.BarValue (FK) ==> (Foo.FooValue)

We then assert a 0 count for all tables