Replaced SupportsTemporalTables flag with CheckSupportsTemporalTables method
Replaced the SupportsTemporalTables flag with an awaitable method that receives the DbConnection arg and
return a boolean indicating whether or not the underlying DB supports Temporal Tables.
Adjusted all IDbAdapter implementations and used the (incorrect) existing one for SQL server.

The CheckSupportsTemporalTables  method has a default implementation to not break code
for users who have written their own IDbAdapter.